Added `RandomReadFile` to provide a cross-platform random-read.
This class currently does not have any tests verifying its behaviour.

It is basically a read-only subset of `RandomAccessFile` from `dart:io` that has a side implementation to work on the web.

This is being introducted here as this library has no buisness being non-js.

All file classes can eventually be expected to use this class and get rid of their dependency on `dart:io` eventually.