Added methds to read ints and floats in big-endian order.
This can be essential for some formats that are stored in that form.

Note that these may evolve in the future to have an argument that accepts the endianness.

The existing, non-suffixed methods will continue to use the host endianness, but may be deprecrated in the future in favour of these newer methods.